docs: add haddock documentation
Fixes #15
